fix(memory): keep extra-file casing policy private
