+++ packages/memory-host-sdk/src/host/explicit-extra-markdown.ts
@@ -1,0 +1,8 @@
+export function isExplicitExtraMarkdownFilePath(
+  filePath: string,
+  platform: NodeJS.Platform = process.platform,
+): boolean {
+  return (
+    filePath.endsWith(".md") || (platform === "win32" && filePath.toLowerCase().endsWith(".md"))
+  );
+}
